requirements.txt
alembic==1.6.2
click==8.0.0
Flask==2.0.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.0
Jinja2==3.0.0
Mako==1.1.4
MarkupSafe==2.0.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==2.0.1
